+++ config/config.go
@@ -13,3 +13,7 @@ type Config struct {
 // to proxy to the service.
 var DefaultTimeout = 5 * time.Minute
 
+// DefaultStopAfter is the default duration of inactivity after which
+// a command will be stopped.
+// The default is zero (the command will not be stopped).
+var DefaultStopAfter = time.Duration(0)